Issue #3987 : common jwk retriever for all sso features that use JWKs
